requirements.txt
fastapi
uvicorn
pymupdf
farm-haystack[faiss]==1.17.2
pytest
